Initial functional implementation
The resource is quite simple. It just calls the send
message endpoint with the right parameters, as documented
here: https://pushover.net/api

To render the message we use `eval` which is prone to
shell injections. But not a big deal: Whoever is running this
resource is able to run any container and any command in the
runners anyway.[1]

The curl returns a json, and it is IMPORTANT that we capture it.
Otherwise smuggler will ignore the versions file. We capture
it and put it in the metadata.

https://github.com/redfactorlabs/concourse-smuggler-resource/issues/19